backend/requirements.txt
python-dotenv
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
google-auth
requests